// apps/admin/src/api/generated/model/participantListDtoOutput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */
import type { ParticipantListDtoOutputItemsItem } from "./participantListDtoOutputItemsItem";

export type ParticipantListDtoOutput = {
  /** @maxItems 500 */
  items: ParticipantListDtoOutputItemsItem[];
};